* Splitted common sync logic into a basesync class for reuse. * Added fullsync command (with configurable sleep interval and page offset) * Added support for netscape cookie format
